fix: handled proper authentication and login page removerd unwanted login/api updated authOptions and middleware
