--- app/api/auth/login/route.ts
@@ -1,53 +1,0 @@
-import { NextResponse } from "next/server";
-import bcrypt from "bcrypt";
-import { query } from "@/lib/db";
-
-export async function POST(req: Request) {
-  try {
-    const { email, password } = await req.json();
-
-    // 1️⃣ Validate input
-    if (!email || !password) {
-      return NextResponse.json(
-        { error: "Email and password are required" },
-        { status: 400 }
-      );
-    }
-
-    // 2️⃣ Find user
-    const result = await query("SELECT * FROM users WHERE email = $1", [email]);
-    if (result.rowCount === 0) {
-      return NextResponse.json(
-        { error: "Invalid credentials" },
-        { status: 401 }
-      );
-    }
-
-    const user = result.rows[0];
-
-    // 3️⃣ Compare password
-    const validPassword = await bcrypt.compare(password, user.password_hash);
-    if (!validPassword) {
-      return NextResponse.json(
-        { error: "Invalid credentials" },
-        { status: 401 }
-      );
-    }
-
-    // 4️⃣ Return success — NextAuth will handle session creation
-    return NextResponse.json({
-      message: "Login successful",
-      user: {
-        id: user.id,
-        email: user.email,
-        full_name: user.full_name,
-      },
-    });
-  } catch (error: any) {
-    console.error("Login error:", error);
-    return NextResponse.json(
-      { error: "Internal Server Error" },
-      { status: 500 }
-    );
-  }
-}
