+++ app/login/page.tsx
@@ -1,90 +1,20 @@
-"use client";
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
+import { authOptions } from "@/lib/authOptions";
+import LoginForm from "@/components/LoginForm";
 
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-import { signIn } from "next-auth/react";
+export default async function LoginPage() {
+  const session = await getServerSession(authOptions);
 
-export default function LoginPage() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [errorMsg, setErrorMsg] = useState("");
-  const router = useRouter();
-
-  async function handleLogin(e: React.FormEvent) {
-    e.preventDefault();
-    try {
-      setLoading(true);
-      setErrorMsg("");
-
-      const res = await signIn("credentials", {
-        redirect: false, // handle redirect manually
-        email,
-        password,
-      });
-
-      if (res?.error) {
-        setErrorMsg(res.error);
-      } else if (res?.ok) {
-        router.push("/dashboard");
-      }
-    } catch (err: any) {
-      setErrorMsg(err.message || "Login failed");
-    } finally {
-      setLoading(false);
-    }
+  // 🚀 Instant redirect if user already authenticated
+  if (session) {
+    redirect("/dashboard");
   }
 
   return (
-    <div className="flex flex-col justify-center items-center min-h-screen px-4">
-      <div className="w-full max-w-md shadow-md rounded-xl p-6">
-        <h1 className="text-2xl font-bold mb-6 text-center">
-          Sign in to Petrol Partner
-        </h1>
-
-        {errorMsg && (
-          <p className="text-red-500 text-sm text-center mb-4">{errorMsg}</p>
-        )}
-
-        <form onSubmit={handleLogin}>
-          <input
-            type="email"
-            placeholder="Email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            className="border border-gray-300 rounded-md p-2 w-full mb-3 focus:ring-2 focus:ring-blue-400 outline-none"
-            disabled={loading}
-            required
-          />
-          <input
-            type="password"
-            placeholder="Password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            className="border border-gray-300 rounded-md p-2 w-full mb-4 focus:ring-2 focus:ring-blue-400 outline-none"
-            disabled={loading}
-            required
-          />
-
-          <button
-            type="submit"
-            disabled={loading}
-            className={`w-full py-2 rounded-md text-white font-semibold transition ${
-              loading
-                ? "bg-blue-400 cursor-not-allowed"
-                : "bg-blue-600 hover:bg-blue-700"
-            }`}
-          >
-            {loading ? "Signing in..." : "Sign In"}
-          </button>
-        </form>
-
-        <p className="text-sm text-gray-500 mt-4 text-center">
-          Don’t have an account?{" "}
-          <a href="/signup" className="text-blue-600 hover:underline">
-            Create one
-          </a>
-        </p>
+    <div className="flex justify-center items-center min-h-screen px-4">
+      <div className="w-full max-w-md">
+        <LoginForm />
       </div>
     </div>
   );
