+++ components/LoginForm.tsx
@@ -1,0 +1,86 @@
+"use client";
+
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import { signIn } from "next-auth/react";
+
+export default function LoginForm() {
+  const router = useRouter();
+
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [loading, setLoading] = useState(false);
+  const [errorMsg, setErrorMsg] = useState("");
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setLoading(true);
+    setErrorMsg("");
+
+    const res = await signIn("credentials", {
+      redirect: false,
+      email,
+      password,
+    });
+
+    if (res?.error) {
+      setErrorMsg(res.error);
+      setLoading(false);
+    } else if (res?.ok) {
+      router.push("/dashboard");
+    }
+  };
+
+  return (
+    <div className="shadow-lg rounded-2xl p-6">
+      <h1 className="text-2xl font-semibold text-center mb-6">
+        Sign in to Petrol Partner
+      </h1>
+
+      {errorMsg && (
+        <p className="text-red-500 text-sm text-center mb-4">{errorMsg}</p>
+      )}
+
+      <form onSubmit={handleSubmit}>
+        <input
+          type="email"
+          placeholder="Email"
+          className="border border-gray-300 rounded-md p-2 w-full mb-3 focus:ring-2 focus:ring-blue-400 outline-none"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          disabled={loading}
+          required
+        />
+
+        <input
+          type="password"
+          placeholder="Password"
+          className="border border-gray-300 rounded-md p-2 w-full mb-4 focus:ring-2 focus:ring-blue-400 outline-none"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          disabled={loading}
+          required
+        />
+
+        <button
+          type="submit"
+          disabled={loading}
+          className={`w-full py-2 rounded-md text-white font-semibold transition ${
+            loading
+              ? "bg-blue-400 cursor-not-allowed"
+              : "bg-blue-600 hover:bg-blue-700"
+          }`}
+        >
+          {loading ? "Signing in..." : "Sign In"}
+        </button>
+      </form>
+
+      <p className="text-sm text-gray-500 mt-4 text-center">
+        Don’t have an account?{" "}
+        <a href="/signup" className="text-blue-600 hover:underline">
+          Create one
+        </a>
+      </p>
+    </div>
+  );
+}
